requirements.txt
opentelemetry-instrumentation==0.36b0
opentelemetry-distro==0.36b0

opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-django==0.36b0
opentelemetry-instrumentation-mysql==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
#Django==2.2.28
#Django==3.1.14
Django

mysqlclient
mysql-connector-python
gunicorn
